refactor(fcp): centralize n2n feed metadata
Move node metadata handling into a shared helper to remove duplication while keeping FCP headers consistent.
