+++ src/main/java/network/crypta/clients/fcp/TextFeedMessage.java
@@ -1,6 +1,7 @@
 package network.crypta.clients.fcp;
 
 import java.nio.charset.StandardCharsets;
+import network.crypta.support.SimpleFieldSet;
 import network.crypta.support.api.Bucket;
 import network.crypta.support.io.ArrayBucket;
 import network.crypta.support.io.NullBucket;
@@ -8,11 +9,11 @@ import network.crypta.support.io.NullBucket;
 /**
  * Feed message that carries a textual node‑to‑node notification with an optional extra body.
  *
- * <p>This class extends {@link N2NFeedMessage} and therefore behaves like a regular {@link
- * FeedMessage} enriched with information about the originating node and transmission timestamps. In
- * addition to the standard {@code "Text"} payload created by the superclass, it attaches a second
- * payload bucket named {@code "MessageText"}, which can be used to transport an auxiliary textual
- * fragment such as a detailed description, log excerpt, or peer‑specific note.
+ * <p>This class behaves like a regular {@link FeedMessage} enriched with information about the
+ * originating node and transmission timestamps. In addition to the standard {@code "Text"} payload
+ * created by the superclass, it attaches a second payload bucket named {@code "MessageText"}, which
+ * can be used to transport an auxiliary textual fragment such as a detailed description, log
+ * excerpt, or peer‑specific note.
  *
  * <p>The primary constructor accepts all fields at once, so instances are fully initialized and
  * effectively immutable after construction. Callers typically create a {@code TextFeedMessage} on a
@@ -24,11 +25,10 @@ import network.crypta.support.io.NullBucket;
  * <p>Thread‑safety: instances are not designed for concurrent mutation but are safe for read‑only
  * access from a single thread once constructed.
  *
- * @see N2NFeedMessage
  * @see FeedMessage
  * @see network.crypta.support.SimpleFieldSet
  */
-public class TextFeedMessage extends N2NFeedMessage {
+public class TextFeedMessage extends FeedMessage {
 
   /**
    * Protocol name of this message type used on the FCP wire.
@@ -40,19 +40,21 @@ public class TextFeedMessage extends N2NFeedMessage {
    */
   public static final String NAME = "TextFeed";
 
+  private final N2NFeedMetadata metadata;
+
   /**
    * Create a new node‑to‑node feed message that carries textual content and an optional extra body.
    *
    * <p>The parameters that describe the header, short summary, and main body are forwarded
-   * unchanged to the {@link N2NFeedMessage} constructor, which in turn populates the {@code "Text"}
-   * bucket and base header fields. This constructor additionally creates a {@link Bucket} for the
-   * {@code messageText} argument and stores it under the key {@code "MessageText"} in the internal
-   * bucket map. When {@code messageText} is {@code null}, a {@link NullBucket} of zero length is
-   * used so that the field set still exposes a {@code MessageTextLength} entry consistent with the
-   * data‑carrying contract.
+   * unchanged to {@link FeedMessage}, which populates the {@code "Text"} bucket and base header
+   * fields. This constructor additionally creates a {@link Bucket} for the {@code messageText}
+   * argument and stores it under the key {@code "MessageText"} in the internal bucket map. When
+   * {@code messageText} is {@code null}, a {@link NullBucket} of zero length is used so that the
+   * field set still exposes a {@code MessageTextLength} entry consistent with the data‑carrying
+   * contract.
    *
    * <p>The resulting instance is ready to be serialized and sent to an FCP client. No further
-   * mutation of bucket contents is expected once construction completes.
+   * mutation of bucket contents is expected once construction is completed.
    *
    * @param params bundled feed metadata describing the header, body text, priority, update time,
    *     source node name, and timing information for this entry.
@@ -61,7 +63,15 @@ public class TextFeedMessage extends N2NFeedMessage {
    *     ArrayBucket}.
    */
   public TextFeedMessage(N2NFeedMessageParams params, String messageText) {
-    super(params);
+    super(
+        params.header(),
+        params.shortText(),
+        params.text(),
+        params.priorityClass(),
+        params.updatedTime());
+    this.metadata =
+        new N2NFeedMetadata(
+            params.sourceNodeName(), params.composed(), params.sent(), params.received());
     final Bucket messageTextBucket;
     if (messageText != null) {
       messageTextBucket = new ArrayBucket(messageText.getBytes(StandardCharsets.UTF_8));
@@ -69,6 +79,19 @@ public class TextFeedMessage extends N2NFeedMessage {
       messageTextBucket = new NullBucket();
     }
     buckets.put("MessageText", messageTextBucket);
+  }
+
+  /**
+   * Build the field set describing this text feed entry.
+   *
+   * <p>The base feed metadata is augmented with node-to-node timing information before being
+   * returned to callers.
+   *
+   * @return a new {@link SimpleFieldSet} that includes node metadata
+   */
+  @Override
+  public SimpleFieldSet getFieldSet() {
+    return metadata.applyTo(super.getFieldSet());
   }
 
   /**
